+++ common/model/src/test/java/is/codion/common/model/loadtest/DefaultLoadTestTest.java
@@ -4,6 +4,7 @@
 package is.codion.common.model.loadtest;
 
 import is.codion.common.model.loadtest.LoadTest.Scenario;
+import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.common.user.User;
 
 import org.junit.jupiter.api.Test;
@@ -22,20 +23,17 @@ public final class DefaultLoadTestTest {
   private static final User UNIT_TEST_USER =
           User.parse(System.getProperty("codion.test.user", "scott:tiger"));
 
-  private static final Scenario<Object> SCENARIO = new AbstractScenario<Object>("test") {
+  private static final Scenario<Object> SCENARIO = Scenario.builder(new Performer<Object>() {
     int counter = 0;
     @Override
-    protected void perform(Object application) throws Exception {
+    public void perform(Object application) throws Exception {
       if (counter++ % 2 == 0) {
         throw new Exception();
       }
     }
-  };
+  }).build();
 
-  private static final Scenario<Object> SCENARIO_II = new AbstractScenario<Object>("testII") {
-    @Override
-    protected void perform(Object application) {}
-  };
+  private static final Scenario<Object> SCENARIO_II = Scenario.builder(application -> {}).build();
 
   @Test
   void unknownScenario() {