+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/testing/scenarios/LogoutLogin.java
@@ -3,15 +3,15 @@
  */
 package is.codion.framework.demos.chinook.testing.scenarios;
 
-import is.codion.common.model.loadtest.AbstractScenario;
+import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.db.EntityConnectionProvider;
 
 import static is.codion.framework.demos.chinook.testing.scenarios.LoadTestUtil.RANDOM;
 
-public final class LogoutLogin extends AbstractScenario<EntityConnectionProvider> {
+public final class LogoutLogin implements Performer<EntityConnectionProvider> {
 
   @Override
-  protected void perform(EntityConnectionProvider connectionProvider) {
+  public void perform(EntityConnectionProvider connectionProvider) {
     try {
       connectionProvider.close();
       Thread.sleep(RANDOM.nextInt(1500));