+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/testing/scenarios/RaisePrices.java
@@ -3,7 +3,7 @@
  */
 package is.codion.framework.demos.chinook.testing.scenarios;
 
-import is.codion.common.model.loadtest.AbstractScenario;
+import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.db.EntityConnection;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.demos.chinook.domain.Chinook.Album;
@@ -19,12 +19,12 @@ import java.util.List;
 import static is.codion.framework.db.EntityConnection.Select.where;
 import static is.codion.framework.demos.chinook.testing.scenarios.LoadTestUtil.randomArtistId;
 
-public final class RaisePrices extends AbstractScenario<EntityConnectionProvider> {
+public final class RaisePrices implements Performer<EntityConnectionProvider> {
 
   private static final BigDecimal PRICE_INCREASE = BigDecimal.valueOf(0.01);
 
   @Override
-  protected void perform(EntityConnectionProvider connectionProvider) throws Exception {
+  public void perform(EntityConnectionProvider connectionProvider) throws Exception {
     EntityConnection connection = connectionProvider.connection();
     Entity artist = connection.selectSingle(Artist.ID.equalTo(randomArtistId()));
     List<Entity> albums = connection.select(where(Album.ARTIST_FK.equalTo(artist))