+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/testing/scenarios/RandomPlaylist.java
@@ -4,7 +4,7 @@
 package is.codion.framework.demos.chinook.testing.scenarios;
 
 import is.codion.common.db.exception.DatabaseException;
-import is.codion.common.model.loadtest.AbstractScenario;
+import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.db.EntityConnection;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.demos.chinook.domain.Chinook.Genre;
@@ -20,14 +20,14 @@ import java.util.UUID;
 import static is.codion.framework.demos.chinook.testing.scenarios.LoadTestUtil.RANDOM;
 import static java.util.Arrays.asList;
 
-public final class RandomPlaylist extends AbstractScenario<EntityConnectionProvider> {
+public final class RandomPlaylist implements Performer<EntityConnectionProvider> {
 
   private static final String PLAYLIST_NAME = "Random playlist";
   private static final Collection<String> GENRES =
           asList("Alternative", "Rock", "Metal", "Heavy Metal", "Pop");
 
   @Override
-  protected void perform(EntityConnectionProvider connectionProvider) throws Exception {
+  public void perform(EntityConnectionProvider connectionProvider) throws Exception {
     EntityConnection connection = connectionProvider.connection();
     List<Entity> playlistGenres = connection.select(Genre.NAME.in(GENRES));
     RandomPlaylistParameters parameters = new RandomPlaylistParameters(PLAYLIST_NAME + " " + UUID.randomUUID(),