+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/testing/scenarios/ViewInvoice.java
@@ -3,7 +3,7 @@
  */
 package is.codion.framework.demos.chinook.testing.scenarios;
 
-import is.codion.common.model.loadtest.AbstractScenario;
+import is.codion.common.model.loadtest.LoadTest.Scenario.Performer;
 import is.codion.framework.db.EntityConnection;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.demos.chinook.domain.Chinook.Customer;
@@ -16,10 +16,10 @@ import java.util.List;
 import static is.codion.framework.demos.chinook.testing.scenarios.LoadTestUtil.RANDOM;
 import static is.codion.framework.demos.chinook.testing.scenarios.LoadTestUtil.randomCustomerId;
 
-public final class ViewInvoice extends AbstractScenario<EntityConnectionProvider> {
+public final class ViewInvoice implements Performer<EntityConnectionProvider> {
 
   @Override
-  protected void perform(EntityConnectionProvider connectionProvider) throws Exception {
+  public void perform(EntityConnectionProvider connectionProvider) throws Exception {
     EntityConnection connection = connectionProvider.connection();
     Entity customer = connection.selectSingle(Customer.ID.equalTo(randomCustomerId()));
     List<Long> invoiceIds = connection.select(Invoice.ID, Invoice.CUSTOMER_FK.equalTo(customer));
@@ -27,10 +27,5 @@ public final class ViewInvoice extends AbstractScenario<EntityConnectionProvider
       Entity invoice = connection.selectSingle(Invoice.ID.equalTo(invoiceIds.get(RANDOM.nextInt(invoiceIds.size()))));
       connection.select(InvoiceLine.INVOICE_FK.equalTo(invoice));
     }
-  }
-
-  @Override
-  public int defaultWeight() {
-    return 10;
   }
 }