+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/InsertDepartment.java
@@ -6,15 +6,15 @@ package is.codion.framework.demos.employees.testing.scenarios;
 import is.codion.framework.demos.employees.domain.Employees.Department;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
 import is.codion.swing.framework.model.SwingEntityModel;
-import is.codion.swing.framework.model.tools.loadtest.AbstractEntityScenario;
+import is.codion.swing.framework.model.tools.loadtest.AbstractEntityPerformer;
 
 import static is.codion.framework.domain.entity.test.EntityTestUtil.createRandomEntity;
 
 // tag::loadTest[]
-public final class InsertDepartment extends AbstractEntityScenario<EmployeesAppModel> {
+public final class InsertDepartment extends AbstractEntityPerformer<EmployeesAppModel> {
 
   @Override
-  protected void perform(EmployeesAppModel application) throws Exception {
+  public void perform(EmployeesAppModel application) throws Exception {
     SwingEntityModel departmentModel = application.entityModel(Department.TYPE);
     departmentModel.editModel().set(createRandomEntity(application.entities(), Department.TYPE, null));
     departmentModel.editModel().insert();