+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/InsertEmployee.java
@@ -9,7 +9,7 @@ import is.codion.framework.demos.employees.model.EmployeesAppModel;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
 import is.codion.swing.framework.model.SwingEntityModel;
-import is.codion.swing.framework.model.tools.loadtest.AbstractEntityScenario;
+import is.codion.swing.framework.model.tools.loadtest.AbstractEntityPerformer;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -17,10 +17,10 @@ import java.util.Map;
 import static is.codion.framework.domain.entity.test.EntityTestUtil.createRandomEntity;
 
 // tag::loadTest[]
-public final class InsertEmployee extends AbstractEntityScenario<EmployeesAppModel> {
+public final class InsertEmployee extends AbstractEntityPerformer<EmployeesAppModel> {
 
   @Override
-  protected void perform(EmployeesAppModel application) throws Exception {
+  public void perform(EmployeesAppModel application) throws Exception {
     SwingEntityModel departmentModel = application.entityModel(Department.TYPE);
     selectRandomRow(departmentModel.tableModel());
     SwingEntityModel employeeModel = departmentModel.detailModel(Employee.TYPE);
@@ -28,11 +28,6 @@ public final class InsertEmployee extends AbstractEntityScenario<EmployeesAppMod
     foreignKeyEntities.put(Employee.DEPARTMENT_FK, departmentModel.tableModel().selectionModel().getSelectedItem());
     employeeModel.editModel().set(createRandomEntity(application.entities(), Employee.TYPE, foreignKeyEntities));
     employeeModel.editModel().insert();
-  }
-
-  @Override
-  public int defaultWeight() {
-    return 3;
   }
 }
 // end::loadTest[]