+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/SelectDepartment.java
@@ -5,19 +5,14 @@ package is.codion.framework.demos.employees.testing.scenarios;
 
 import is.codion.framework.demos.employees.domain.Employees.Department;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
-import is.codion.swing.framework.model.tools.loadtest.AbstractEntityScenario;
+import is.codion.swing.framework.model.tools.loadtest.AbstractEntityPerformer;
 
 // tag::loadTest[]
-public final class SelectDepartment extends AbstractEntityScenario<EmployeesAppModel> {
+public final class SelectDepartment extends AbstractEntityPerformer<EmployeesAppModel> {
 
   @Override
-  protected void perform(EmployeesAppModel application) {
+  public void perform(EmployeesAppModel application) {
     selectRandomRow(application.entityModel(Department.TYPE).tableModel());
-  }
-
-  @Override
-  public int defaultWeight() {
-    return 10;
   }
 }
 // end::loadTest[]