+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/scenarios/UpdateEmployee.java
@@ -9,19 +9,19 @@ import is.codion.framework.demos.employees.domain.Employees.Employee;
 import is.codion.framework.demos.employees.model.EmployeesAppModel;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.swing.framework.model.SwingEntityModel;
-import is.codion.swing.framework.model.tools.loadtest.AbstractEntityScenario;
+import is.codion.swing.framework.model.tools.loadtest.AbstractEntityPerformer;
 
 import java.util.Random;
 
 import static is.codion.framework.domain.entity.test.EntityTestUtil.randomize;
 
 // tag::loadTest[]
-public final class UpdateEmployee extends AbstractEntityScenario<EmployeesAppModel> {
+public final class UpdateEmployee extends AbstractEntityPerformer<EmployeesAppModel> {
 
   private final Random random = new Random();
 
   @Override
-  protected void perform(EmployeesAppModel application) throws Exception {
+  public void perform(EmployeesAppModel application) throws Exception {
     SwingEntityModel departmentModel = application.entityModel(Department.TYPE);
     selectRandomRow(departmentModel.tableModel());
     SwingEntityModel employeeModel = departmentModel.detailModel(Employee.TYPE);
@@ -49,11 +49,6 @@ public final class UpdateEmployee extends AbstractEntityScenario<EmployeesAppMod
         }
       }
     }
-  }
-
-  @Override
-  public int defaultWeight() {
-    return 5;
   }
 }
 // end::loadTest[]