+++ demos/manual/src/main/java/is/codion/framework/demos/manual/store/test/StoreLoadTest.java
@@ -4,6 +4,7 @@
 package is.codion.framework.demos.manual.store.test;
 
 import is.codion.common.model.loadtest.LoadTest;
+import is.codion.common.model.loadtest.LoadTest.Scenario;
 import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.rmi.RemoteEntityConnectionProvider;
@@ -13,7 +14,7 @@ import is.codion.framework.demos.manual.store.model.StoreApplicationModel;
 import is.codion.swing.common.model.tools.loadtest.LoadTestModel;
 import is.codion.swing.common.ui.tools.loadtest.LoadTestPanel;
 import is.codion.swing.framework.model.SwingEntityModel;
-import is.codion.swing.framework.model.tools.loadtest.AbstractEntityScenario;
+import is.codion.swing.framework.model.tools.loadtest.AbstractEntityPerformer;
 
 import java.util.function.Function;
 
@@ -37,11 +38,11 @@ public class StoreLoadTest {
     }
   }
 
-  private static class StoreUsageScenario extends
-          AbstractEntityScenario<StoreApplicationModel> {
+  private static class StoreScenarioPerformer extends
+          AbstractEntityPerformer<StoreApplicationModel> {
 
     @Override
-    protected void perform(StoreApplicationModel application) {
+    public void perform(StoreApplicationModel application) {
       SwingEntityModel customerModel = application.entityModel(Customer.TYPE);
       customerModel.tableModel().refresh();
       selectRandomRow(customerModel.tableModel());
@@ -53,7 +54,7 @@ public class StoreLoadTest {
             LoadTest.builder(new StoreApplicationModelFactory(),
                             application -> application.connectionProvider().close())
                     .user(User.parse("scott:tiger"))
-                    .scenarios(singletonList(new StoreUsageScenario()))
+                    .scenarios(singletonList(Scenario.builder(new StoreScenarioPerformer()).build()))
                     .titleFactory(model -> "Store LoadTest - " + EntityConnectionProvider.CLIENT_CONNECTION_TYPE.get())
                     .build();
     new LoadTestPanel<>(LoadTestModel.loadTestModel(loadTest)).run();