+++ demos/petstore/src/main/java/is/codion/framework/demos/petstore/testing/PetstoreLoadTest.java
@@ -4,6 +4,7 @@
 package is.codion.framework.demos.petstore.testing;
 
 import is.codion.common.model.loadtest.LoadTest;
+import is.codion.common.model.loadtest.LoadTest.Scenario;
 import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.demos.petstore.domain.Petstore;
@@ -11,7 +12,7 @@ import is.codion.framework.demos.petstore.model.PetstoreAppModel;
 import is.codion.swing.common.model.tools.loadtest.LoadTestModel;
 import is.codion.swing.common.ui.tools.loadtest.LoadTestPanel;
 import is.codion.swing.framework.model.SwingEntityModel;
-import is.codion.swing.framework.model.tools.loadtest.AbstractEntityScenario;
+import is.codion.swing.framework.model.tools.loadtest.AbstractEntityPerformer;
 
 import java.util.function.Function;
 
@@ -44,14 +45,10 @@ public final class PetstoreLoadTest {
     }
   }
 
-  private static final class PetstoreUsageScenario extends AbstractEntityScenario<PetstoreAppModel> {
-
-    private PetstoreUsageScenario() {
-      super("selectRecords");
-    }
+  private static final class PetstoreUsageScenario extends AbstractEntityPerformer<PetstoreAppModel> {
 
     @Override
-    protected void perform(PetstoreAppModel application) throws Exception {
+    public void perform(PetstoreAppModel application) throws Exception {
       SwingEntityModel categoryModel = application.entityModels().iterator().next();
       categoryModel.tableModel().selectionModel().clearSelection();
       categoryModel.tableModel().refresh();
@@ -66,7 +63,9 @@ public final class PetstoreLoadTest {
             LoadTest.builder(new PetstoreAppModelFactory(),
                             application -> application.connectionProvider().close())
                     .user(UNIT_TEST_USER)
-                    .scenarios(singletonList(new PetstoreUsageScenario()))
+                    .scenarios(singletonList(Scenario.builder(new PetstoreUsageScenario())
+                            .name("selectRecords")
+                            .build()))
                     .titleFactory(model -> "Petstore LoadTest - " + EntityConnectionProvider.CLIENT_CONNECTION_TYPE.get())
                     .build();
     new LoadTestPanel<>(LoadTestModel.loadTestModel(loadTest)).run();